Move config logic to pkg
