+++ pkg/cmd/pulumi/config/io_test.go
@@ -507,7 +507,7 @@ func TestStackEnvConfig(t *testing.T) {
 	assert.Nil(t, cfg.Config)
 	cfg.Config = config.Map{}
 
-	err = workspace.ApplyProjectConfig(
+	err = pkgWorkspace.ApplyProjectConfig(
 		ctx, "mystack", &project, cfg.Environment, cfg.Config, config.NopEncrypter, config.NopDecrypter,
 	)
 	require.NoError(t, err)
