+++ pkg/cmd/pulumi/operations/up.go
@@ -196,12 +196,12 @@ func NewUpCmd() *cobra.Command {
 		stackName := s.Ref().Name().String()
 		if skipConfigValidation {
 			// Still apply project config defaults onto the stack config, but skip validation.
-			if configErr := workspace.ApplyProjectConfig(
+			if configErr := pkgWorkspace.ApplyProjectConfig(
 				ctx, stackName, proj, cfg.Environment, cfg.Config, encrypter, decrypter); configErr != nil {
 				return fmt.Errorf("applying stack config: %w", configErr)
 			}
 		} else {
-			configErr := workspace.ValidateStackConfigAndApplyProjectConfig(
+			configErr := pkgWorkspace.ValidateStackConfigAndApplyProjectConfig(
 				ctx,
 				stackName,
 				proj,
@@ -491,12 +491,12 @@ func NewUpCmd() *cobra.Command {
 		stackName := s.Ref().String()
 		if skipConfigValidation {
 			// Still apply project config defaults onto the stack config, but skip validation.
-			if configErr := workspace.ApplyProjectConfig(
+			if configErr := pkgWorkspace.ApplyProjectConfig(
 				ctx, stackName, proj, cfg.Environment, cfg.Config, encrypter, decrypter); configErr != nil {
 				return fmt.Errorf("applying stack config: %w", configErr)
 			}
 		} else {
-			configErr := workspace.ValidateStackConfigAndApplyProjectConfig(
+			configErr := pkgWorkspace.ValidateStackConfigAndApplyProjectConfig(
 				ctx,
 				stackName,
 				proj,
