+++ sdk/go/common/workspace/project_test.go
@@ -17,14 +17,11 @@ package workspace
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 	"os"
 	"testing"
 
 	"github.com/pulumi/pulumi/sdk/v3/go/common/diag"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/encoding"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/esc"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/resource/config"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/slice"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/testing/diagtest"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/tokens"
@@ -556,258 +553,6 @@ config:
 	assert.Nil(t, secretString.Items)
 }
 
-func getConfigValue(t *testing.T, stackConfig config.Map, key string) string {
-	parsedKey, err := config.ParseKey(key)
-	require.NoErrorf(t, err, "There should be no error parsing the config key '%v'", key)
-	configValue, foundValue := stackConfig[parsedKey]
-	assert.Truef(t, foundValue, "Couldn't find a value for config key %v", key)
-	value, valueError := configValue.Value(config.NopDecrypter)
-	require.NoErrorf(t, valueError, "Error while getting the value for key %v", key)
-	return value
-}
-
-func getConfigValueUnmarshalled(t *testing.T, stackConfig config.Map, key string) any {
-	parsedKey, err := config.ParseKey(key)
-	require.NoErrorf(t, err, "There should be no error parsing the config key '%v'", key)
-	configValue, foundValue := stackConfig[parsedKey]
-	assert.Truef(t, foundValue, "Couldn't find a value for config key %v", key)
-	valueJSON, valueError := configValue.Value(config.NopDecrypter)
-	require.NoErrorf(t, valueError, "Error while getting the value for key %v", key)
-	var value any
-	err = json.Unmarshal([]byte(valueJSON), &value)
-	require.NoErrorf(t, err, "Error while unmarshalling value for key %v", key)
-	return value
-}
-
-func TestStackConfigIsInheritedFromProjectConfig(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  instanceSize: t3.micro
-  instanceCount: 20
-  protect: true`
-
-	projectStackYaml := `
-config:
-  test:instanceSize: t4.large`
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	require.NoError(t, configError, "Config override should be valid")
-
-	require.Len(t, stack.Config, 3, "Stack config now has three values")
-	// value of instanceSize is overwritten from the stack
-	assert.Equal(t, "t4.large", getConfigValue(t, stack.Config, "test:instanceSize"))
-	// instanceCount and protect are inherited from the project
-	assert.Equal(t, "20", getConfigValue(t, stack.Config, "test:instanceCount"))
-	assert.Equal(t, "true", getConfigValue(t, stack.Config, "test:protect"))
-}
-
-func TestNamespacedConfigValuesAreInheritedCorrectly(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  aws:region: us-west-1
-  pulumi:disable-default-providers: ["*"]
-  instanceSize: t3.micro`
-
-	projectStackYaml := `
-config:
-  test:instanceSize: t4.large`
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	require.NoError(t, configError, "Config override should be valid")
-	require.Len(t, stack.Config, 3, "Stack config now has three values")
-	// value of instanceSize is overwritten from the stack
-	assert.Equal(t, "t4.large", getConfigValue(t, stack.Config, "test:instanceSize"))
-	// aws:region is namespaced and is inherited from the project
-	assert.Equal(t, "us-west-1", getConfigValue(t, stack.Config, "aws:region"))
-	assert.Equal(t, "[\"*\"]", getConfigValue(t, stack.Config, "pulumi:disable-default-providers"))
-	assert.Equal(t, []any{"*"}, getConfigValueUnmarshalled(t, stack.Config, "pulumi:disable-default-providers"))
-}
-
-func TestLoadingStackConfigWithoutNamespacingTheProject(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  aws:region: us-west-1
-  instanceSize: t3.micro`
-
-	projectStackYaml := `
-config:
-  instanceSize: t4.large`
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	require.NoError(t, configError, "Config override should be valid")
-
-	require.Len(t, stack.Config, 2, "Stack config now has three values")
-	// value of instanceSize is overwritten from the stack
-	assert.Equal(t, "t4.large", getConfigValue(t, stack.Config, "test:instanceSize"))
-	// aws:region is namespaced and is inherited from the project
-	assert.Equal(t, "us-west-1", getConfigValue(t, stack.Config, "aws:region"))
-}
-
-func TestUntypedProjectConfigValuesAreNotValidated(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  instanceSize: t3.micro
-  aws:region: us-west-1`
-
-	projectStackYaml := `
-config:
-  instanceSize: 9999
-  aws:region: 42`
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	require.NoError(t, configError, "Config override should be valid")
-	require.Len(t, stack.Config, 2, "Stack config now has three values")
-	// value of instanceSize is overwritten from the stack
-	assert.Equal(t, "9999", getConfigValue(t, stack.Config, "test:instanceSize"))
-	assert.Equal(t, "42", getConfigValue(t, stack.Config, "aws:region"))
-}
-
-func TestUntypedProjectConfigValuesWithOnlyDefaultOrOnlyValue(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  instanceSize:
-    default: t3.micro
-  region:
-    value: us-west-1`
-
-	projectStackYaml := `
-config:
-  aws:answer: 42`
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	require.NoError(t, configError, "Config override should be valid")
-	require.Len(t, stack.Config, 3, "Stack config now has three values")
-	// value of instanceSize is overwritten from the stack
-	assert.Equal(t, "t3.micro", getConfigValue(t, stack.Config, "test:instanceSize"))
-	assert.Equal(t, "us-west-1", getConfigValue(t, stack.Config, "test:region"))
-	assert.Equal(t, "42", getConfigValue(t, stack.Config, "aws:answer"))
-}
-
-func TestUntypedStackConfigValuesDoNeedProjectDeclaration(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  createVpc: true`
-
-	projectStackYaml := `
-config:
-  instanceSize: 42`
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	require.NoError(t, configError, "Config override should be valid")
-	require.Len(t, stack.Config, 2, "Stack config now has three values")
-	// value of instanceSize is overwritten from the stack
-	assert.Equal(t, "42", getConfigValue(t, stack.Config, "test:instanceSize"))
-	assert.Equal(t, "true", getConfigValue(t, stack.Config, "test:createVpc"))
-}
-
 func TestNamespacedProjectConfigShouldNotBeExplicitlyTyped(t *testing.T) {
 	t.Parallel()
 	projectYaml := `
@@ -872,83 +617,6 @@ config:
 		"Did you mean to use the 'value' attribute instead of 'default'?")
 }
 
-func TestUntypedProjectConfigObjectValuesPassedDownToStack(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  instanceSize:
-    value:
-      hello: world
-  aws:config:
-    value:
-      region: us-west-1`
-
-	projectStackYaml := `
-config:
-  aws:whatever: 42`
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	require.NoError(t, configError, "Config override should be valid")
-	require.Len(t, stack.Config, 3, "Stack config now has three values")
-	// value of instanceSize is overwritten from the stack
-	assert.Equal(t, "{\"hello\":\"world\"}", getConfigValue(t, stack.Config, "test:instanceSize"))
-	assert.Equal(t, "{\"region\":\"us-west-1\"}", getConfigValue(t, stack.Config, "aws:config"))
-	assert.Equal(t, "42", getConfigValue(t, stack.Config, "aws:whatever"))
-}
-
-func TestStackConfigErrorsWhenStackValueIsNotCorrectlyTyped(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  values:
-    type: array
-    items:
-      type: string
-    default: [value]`
-
-	projectStackYaml := `
-config:
-  test:values: someValue
-`
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	assert.ErrorContains(t, configError, "Stack 'dev' with configuration key 'values' must be of type 'array<string>'")
-}
-
 func TestLoadingConfigIsRewrittenToStackConfigDir(t *testing.T) {
 	t.Parallel()
 	projectYaml := `
@@ -990,328 +658,6 @@ config:
 	require.NotNil(t, project, "The project can be loaded correctly")
 	assert.Equal(t, "./some/other/path", project.StackConfigDir)
 	require.Len(t, project.Config, 1, "there is one config value")
-}
-
-func TestStackConfigIntegerTypeIsCorrectlyValidated(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  importantNumber:
-    type: integer
-`
-
-	projectStackYamlValid := `
-config:
-  test:importantNumber: 20
-`
-
-	projectStackYamlInvalid := `
-config:
-  test:importantNumber: hello
-`
-
-	ctx := t.Context()
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYamlValid)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		ctx,
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	require.NoError(t, configError, "there should no config type error")
-
-	invalidStackConfig, stackError := loadProjectStackFromText(t, sink, project, projectStackYamlInvalid)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError = ValidateStackConfigAndApplyProjectConfig(
-		ctx,
-		"dev",
-		project,
-		esc.Value{},
-		invalidStackConfig.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	assert.ErrorContains(t, configError,
-		"Stack 'dev' with configuration key 'importantNumber' must be of type 'integer'")
-}
-
-func TestStackConfigErrorsWhenMissingStackValueForConfigTypeWithNoDefault(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  values:
-    type: array
-    items:
-      type: string`
-
-	projectStackYaml := ``
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	assert.ErrorContains(t, configError, "Stack 'dev' is missing configuration value 'values'")
-}
-
-func TestStackConfigErrorsWhenMissingTwoStackValueForConfigTypeWithNoDefault(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  another:
-    type: string
-  values:
-    type: array
-    items:
-      type: string`
-
-	projectStackYaml := ``
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	assert.ErrorContains(t, configError, "Stack 'dev' is missing configuration values 'another' and 'values'")
-}
-
-func TestStackConfigErrorsWhenMissingMultipleStackValueForConfigTypeWithNoDefault(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  hello:
-    type: integer
-  values:
-    type: array
-    items:
-      type: string
-  world:
-    type: string`
-
-	projectStackYaml := ``
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	assert.ErrorContains(t, configError, "Stack 'dev' is missing configuration values 'hello', 'values' and 'world'")
-}
-
-func TestStackConfigDoesNotErrorWhenProjectHasNotDefinedConfig(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet`
-
-	projectStackYaml := `
-config:
-  hello: 21
-  world: 42
-  another: 42`
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	assert.Nil(t, configError, "there should not be a config type error")
-}
-
-func TestStackConfigSecretIsCorrectlyValidated(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet
-config:
-  importantNumber:
-    type: integer
-    secret: true
-`
-
-	crypter := config.Base64Crypter
-	encryptedValue, err := crypter.EncryptValue(t.Context(), "20")
-	require.NoError(t, err)
-
-	projectStackYamlValid := fmt.Sprintf(`
-config:
-  test:importantNumber:
-    secure: %s
-`, encryptedValue)
-
-	projectStackYamlInvalid := `
-config:
-  test:importantNumber: 20
-`
-
-	ctx := t.Context()
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYamlValid)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		ctx,
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		crypter,
-		crypter)
-	require.NoError(t, configError, "there should no config type error")
-
-	invalidStackConfig, stackError := loadProjectStackFromText(t, sink, project, projectStackYamlInvalid)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError = ValidateStackConfigAndApplyProjectConfig(
-		ctx,
-		"dev",
-		project,
-		esc.Value{},
-		invalidStackConfig.Config,
-		crypter,
-		crypter)
-	assert.ErrorContains(t, configError,
-		"Stack 'dev' with configuration key 'importantNumber' must be encrypted as it's secret")
-}
-
-//nolint:lll
-func TestEnvironmentMerge(t *testing.T) {
-	t.Parallel()
-
-	projectYAML := `
-name: test
-runtime: nodejs`
-
-	stackYAML := `
-config:
-  test:boolean: true
-  test:number: 42
-  test:string: foo
-  test:array:
-    - first
-    - second
-  test:object:
-    foo: bar
-    object:
-      baz: 42`
-
-	env := esc.NewValue(map[string]esc.Value{
-		"test:boolean": esc.NewValue(false),
-		"test:number":  esc.NewValue(json.Number("42")),
-		"test:string":  esc.NewValue("esc"),
-		"test:array":   esc.NewValue([]esc.Value{esc.NewValue("second"), esc.NewValue("first")}),
-		"test:secret":  esc.NewSecret("hunter2"),
-		"test:object": esc.NewValue(map[string]esc.Value{
-			"boolean": esc.NewValue(true),
-			"number":  esc.NewValue(json.Number("42")),
-			"string":  esc.NewValue("esc"),
-			"array":   esc.NewValue([]esc.Value{esc.NewValue("first"), esc.NewValue("second")}),
-			"object":  esc.NewValue(map[string]esc.Value{"foo": esc.NewValue("bar")}),
-			"foo":     esc.NewValue("qux"),
-		}),
-	})
-
-	project, projectError := loadProjectFromText(t, projectYAML)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, stackYAML)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		env,
-		stack.Config,
-		config.Base64Crypter,
-		config.Base64Crypter)
-	require.NoError(t, configError, "there should not be a config type error")
-
-	secureKeys := stack.Config.SecureKeys()
-	assert.Equal(t, []config.Key{config.MustMakeKey("test", "secret")}, secureKeys)
-
-	m, err := stack.Config.Decrypt(config.Base64Crypter)
-	require.NoError(t, err)
-
-	expected := map[config.Key]string{
-		config.MustMakeKey("test", "array"):   "[\"first\",\"second\"]",
-		config.MustMakeKey("test", "boolean"): "true",
-		config.MustMakeKey("test", "number"):  "42",
-		config.MustMakeKey("test", "object"):  "{\"array\":[\"first\",\"second\"],\"boolean\":true,\"foo\":\"bar\",\"number\":42,\"object\":{\"baz\":42,\"foo\":\"bar\"},\"string\":\"esc\"}",
-		config.MustMakeKey("test", "string"):  "foo",
-		config.MustMakeKey("test", "secret"):  "hunter2",
-	}
-	assert.Equal(t, expected, m)
 }
 
 func TestProjectLoadYAML(t *testing.T) {
@@ -1436,7 +782,8 @@ func TestProjectEditRoundtrip(t *testing.T) {
 					proj.Runtime.Name(),
 					map[string]any{
 						"setting": "test",
-					})
+					},
+				)
 			},
 			expected: "name: test\nruntime:\n  name: python\n  options:\n    setting: test\n",
 		},
@@ -2018,37 +1365,6 @@ runtime: yaml`
 }
 
 // Regression test for https://github.com/pulumi/pulumi/issues/18581, check that we handle uint64's in yaml
-func TestStackConfigUInt64(t *testing.T) {
-	t.Parallel()
-	projectYaml := `
-name: test
-runtime: dotnet`
-
-	projectStackYaml := `
-config:
-  test:instanceSize: 18446744073709551615`
-
-	project, projectError := loadProjectFromText(t, projectYaml)
-	require.NoError(t, projectError, "Shold be able to load the project")
-	var stdout, stderr bytes.Buffer
-	sink := diagtest.MockSink(&stdout, &stderr)
-	stack, stackError := loadProjectStackFromText(t, sink, project, projectStackYaml)
-	assert.Empty(t, stdout)
-	assert.Empty(t, stderr)
-	require.NoError(t, stackError, "Should be able to read the stack")
-	configError := ValidateStackConfigAndApplyProjectConfig(
-		t.Context(),
-		"dev",
-		project,
-		esc.Value{},
-		stack.Config,
-		config.NewPanicCrypter(),
-		config.NewPanicCrypter())
-	require.NoError(t, configError, "Config override should be valid")
-
-	require.Len(t, stack.Config, 1, "Stack config now has three values")
-	assert.Equal(t, "18446744073709551615", getConfigValue(t, stack.Config, "test:instanceSize"))
-}
 
 func TestPackageValueSerialization(t *testing.T) {
 	t.Parallel()
